move build status files to separate folder
